faet #7 : global container 적용 및 header
